Moves `LocalFeedImageDataLoader` and `FeedImageDataStore` to production
